Bug fix createConnectionChain: wrong socket returned
The last socket in the stack of proxies tunnels to the final destination. Therefore the last socket has to be returned by resolve/callback. The current implementation returned the first socket of the stack - instead of the last one. This suggested bug fix solves that. Note: tested only locally - but no unit tests provided. Hope this is OK.