Bug fix createConnectionChain - creating a tunnel (#88)
The current implementation returns a socket to the first proxy in the chain - this is the bug (reproduce: the example in the README does not work). The method "createConnectionChain" should return a socket to the final destination which is tunnelled through the given proxies. Therefore each iteration over the array of proxies must carry forward the socket. This PR is implementing this and hence fixing the bug.